src/main/java/org/search/driver/PersonDriver.java
package org.search.driver;

import java.io.BufferedReader;
import java.io.FileReader;
import java.io.IOException;
import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;

public final class PersonDriver {
    public static List<String> convertFromFile(String fileName) throws IOException {
        List<String> lines = new ArrayList<>();
        try (BufferedReader reader = new BufferedReader(new FileReader(fileName))) {
            lines = reader.lines().collect(Collectors.toList());
        } catch (IOException exception) {
            throw new IOException("There is no correct file");
        }
        return lines;
    }
}
